pyproject.toml
[tool.poetry]
name = "la-poste"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Flask-Cors = "^3.0.10"
api-client = "^1.3.1"
Flask-SQLAlchemy = "^2.5.1"
Flask = "^2.1.2"
celery = "^5.2.6"
python-dotenv = "^0.20.0"


[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.3.0"
pytest-recording = "^0.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
